Auto Conversion to Paper Item Serialization
Any time a vault is opened it will be saved as a paper vault instead of the bukkit vaults. This should make the items more version safe.

Also adds a new vault command which will forcefully convert all player vaults to Paper serialization as a last minute hurray to run just before updating the server